docs: query hook 파일명과 파일 위치 변경
