+++ src/hooks/queries/reaction/usePostReactionQuery.js
@@ -1,14 +1,6 @@
-import { getReactions, postReactions } from 'api/reaction';
-import { useMutation, useQuery, useQueryClient } from 'react-query';
+import { postReactions } from 'api/reaction';
+import { useMutation, useQueryClient } from 'react-query';
 import { errorAlert } from 'utils/errorAlert';
-
-export const useGetReactionQuery = postId => {
-  return useQuery({
-    queryKey: ['reaction', postId],
-    queryFn: () => getReactions(postId),
-    onError: () => errorAlert({ title: '이모지 조회 실패' }),
-  });
-};
 
 export const usePostReactionQuery = postId => {
   const queryClient = useQueryClient();
